+++ test/framework/easystacks/test_easystack_basic.yaml
@@ -3,4 +3,4 @@ software:
     toolchains:
       GCCcore-4.9.3:
         versions:
-            [2.25]
+          [2.25]
